Check that click on mopub interstitial redirect user and notify mopub
Tested scenario:
- Given already prefetch valid bid
- Given MoPub response pointing to Criteo adapter
- When loading the MoPub interstitial
- And waiting for idle state
- And showing the interstitial
- And clicking on it
- Then user is redirected to interstitial URI
- Then MoPub listener is notified for click

JIRA: EE-1179
