+++ mediation/src/androidTest/java/com/criteo/publisher/StubConstantsExt.kt
@@ -1,0 +1,8 @@
+package com.criteo.publisher
+
+import java.net.URI
+
+/**
+ * URI reached when clicking on banner/interstitial Ads delivered by the Casper stub.
+ */
+val STUB_CLICK_URI = URI.create("https://criteo.com/")
